Local/Kaggle = 0.962009/0.96714
Numerous enhancements to reduce memory consumption to bring in more
training data.

+++ src/dataModel.py
@@ -1,11 +1,13 @@
 from numpy import zeros
+import numpy
 
 class TestExample(object):
     def __init__(self, dictionary):
         self.WIDTH = 28
         self.HEIGHT = 28
 
-        self.X = zeros((self.WIDTH, self.HEIGHT))
+        # 2015-05-15 GEL Changed dtype=numpy.uint8 (0, 255) to reduce memory.
+        self.X = zeros((self.WIDTH, self.HEIGHT), dtype=numpy.uint8)
         
         for key in dictionary:
             if key == "label":
